# docs/swagger.yaml
basePath: /
...
host: dev-internal-api.elmodis.com
info:
  contact:
    email: mateusz-gluch@elmodis.com
    name: Elmodis
  title: Internal Scripts API
  version: "1.0"
paths:
...
  /events-summary/data:
    get:
      consumes:
      - application/json
      parameters:
      - default: 367,333
        description: Comma separated list of Asset IDs
        in: query
        name: assets
        required: true
        type: string
      - default: 1708300800:1708387200
        description: Timestamp range in format {startTs}:{endTs}. Timestamps in seconds,
          tz=UTC
        in: query
        name: ts
        type: string
      - default: 144h
        description: Span description string
        in: query
        name: span
        type: string
      - description: Comma separated list of event categories
        example: MACHINE,DATA
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Retrieves script data for asset context
      tags:
      - Data
...
swagger: "2.0"